requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports-abc==0.5
beautifulsoup4==4.6.0
billiard==3.3.0.23
celery==3.1.17
certifi==2017.4.17
chardet==3.0.4
Django==1.9
django-celery==3.1.17
future==0.16.0
futures==3.1.1
idna==2.5
ipaddress==1.0.18
jenkinsapi==0.3.4
Jinja2==2.9.6
kombu==3.0.37
MarkupSafe==1.0
msgpack-python==0.4.8
MySQL-python==1.2.5
ndg-httpsclient==0.4.4
pyasn1==0.4.2
pycrypto==2.6.1
pydisque==0.1.2
pyOpenSSL==17.5.0
python-consul==0.7.2
python-gitlab==0.21.2
pytz==2017.2
PyYAML==3.12
pyzabbix==0.7.4
pyzmq==16.0.2
qcloudapi-sdk-python==2.0.10
qingcloud-sdk==1.2
redis==2.10.5
requests==2.18.1
salt==2016.11.6
singledispatch==3.4.0.3
six==1.10.0
tornado==4.5.1
urllib3==1.21.1
python-crontab==2.3.4
django-crontab==0.7.1
